Revert "WIP Feature/139 pipeline tests"
