+++ .travis.yml
@@ -2,6 +2,7 @@ matrix:
  include:
   - language: python
     python:
+     - "3.6"
      - "3.7"
     install:
      - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
