+++ pyveg/src/image_utils.py
@@ -21,7 +21,34 @@ import matplotlib
 matplotlib.use('PS')
 import matplotlib.pyplot as plt
 
-from .coordinate_utils import get_sub_image_coords
+
+def save_json(out_dict, output_dir, output_filename):
+    """
+    Given a dictionary, save
+    to requested filename -
+    """
+    if not os.path.exists(output_dir):
+        os.makedirs(output_dir, exist_ok=True)
+    output_path = os.path.join(output_dir, output_filename)
+
+    with open(output_path, 'w') as fp:
+        json.dump(out_dict, fp, indent=2)
+
+    #print("Saved json file '{}'".format(output_path))
+
+
+def save_image(image, output_dir, output_filename):
+    """
+    Given a PIL.Image (list of pixel values), save
+    to requested filename - note that the file extension
+    will determine the output file type, can be .png, .tif,
+    probably others...
+    """
+    if not os.path.exists(output_dir):
+        os.makedirs(output_dir, exist_ok=True)
+    output_path = os.path.join(output_dir, output_filename)
+    image.save(output_path)
+    #print("Saved image '{}'".format(output_path))
 
 
 def image_from_array(input_array, output_size=None, sel_val=200):
@@ -213,7 +240,18 @@ def crop_image_npix(input_image, n_pix_x, n_pix_y=None,
     y_parts = int(ysize // n_pix_y)
 
     # if we are given coords, calculate coords for all sub-regions
-    sub_image_coords = get_sub_image_coords(coords, region_size, x_parts, y_parts)
+    sub_image_coords = []
+    if coords and region_size:
+        left_start = coords[0] - region_size/2
+        top_start = coords[1] + region_size/2
+        sub_image_size_x = region_size / x_parts
+        sub_image_size_y = region_size / y_parts
+        for ix in range(x_parts):
+            for iy in range(y_parts):
+                sub_image_coords.append(
+                    (left_start + sub_image_size_x/2 + (ix*sub_image_size_x),
+                     top_start - sub_image_size_y/2 - (iy*sub_image_size_y))
+                )
 
     # now do the actual cropping
     sub_images = []
@@ -305,8 +343,7 @@ def create_gif_from_images(directory_path, output_name, string_in_filename=""):
 
     :param directory_path:  directory where all the files are.
     :param output_name: name to be given to the output gif
-    :param string_in_filename: select only files that containsa particular string,
-                  default is "" which implies all in directory files are selected
+    :param string_in_filename: select only files that containsa particular string, default is "" which implies all in directory files are selected
 
     :return:
     """
@@ -561,11 +598,6 @@ def check_image_ok(rgb_image, black_pix_threshold=0.05):
     if len(img_array.shape) < 3:
         return False
     black = [0,0,0]
-
-    # catch an error where array elements can be zero, rather than [r,g,b] values
-    if len(img_array.shape) < 3:
-        return False
-
     n_black_pix = np.count_nonzero(np.all(img_array == black, axis=2))
 
     if n_black_pix / (img_array.shape[0]*img_array.shape[1]) >= black_pix_threshold:
