+++ pyveg/src/pattern_generation.py
@@ -5,7 +5,6 @@ import os
 import random
 import json
 import numpy as np
-
 from matplotlib import pyplot as plt
 
 class PatternGenerator(object):
