+++ pyveg/tests/test_gee_download.py
@@ -4,19 +4,11 @@ import shutil
 
 import unittest
 
-from pyveg.src.pyveg_pipeline import Sequence
-
-from pyveg.src.processor_modules import (
-    VegetationImageProcessor,
-    NetworkCentralityCalculator,
-    WeatherImageToJSON
-)
-
 coordinates = (27.99,11.29)
 
 # get one data point for the test
-date_range = ['2016-01-01', '2016-01-31']
-time_per_point = "30d"
+date_range = ('2016-01-01', '2016-01-31')
+num_days_per_point = 30
 
 do_network_centrality = True
 
@@ -24,7 +16,7 @@ data_collections = {
     'Copernicus' : {
         'collection_name': 'COPERNICUS/S2',
         'type': 'vegetation',
-        'RGB_bands': ['B4','B3','B2'],
+        'RGB_bands': ('B4','B3','B2'),
         'NIR_band': 'B8',
         'cloudy_pix_flag': 'CLOUDY_PIXEL_PERCENTAGE',
         'do_network_centrality': do_network_centrality
@@ -47,7 +39,7 @@ data_collections = {
         'type': 'weather',
         'precipitation_band': ['precipitationCal'],
     },
-    'ERA5' : {
+    'unsupported' : {
         'collection_name': "ECMWF/ERA5/DAILY",
         'type': 'weather',
         'precipitation_band': ['total_precipitation'],
@@ -56,40 +48,19 @@ data_collections = {
 }
 
 test_out_dir = 'test_out'
-
 @unittest.skipIf(os.environ.get('TRAVIS') == 'true','Skipping this test on Travis CI.')
 def test_get_vegetation():
-    from pyveg.src.download_modules import VegetationDownloader
-    s = Sequence("vegetation")
-    s.set_config(data_collections["Copernicus"])
-    s.output_dir = test_out_dir
-    s.coords = coordinates
-    s.date_range = date_range
-    s.n_sub_images = 1
-    s.time_per_point = time_per_point
-    s += VegetationDownloader()
-    s += VegetationImageProcessor()
-    s += NetworkCentralityCalculator()
-    s.configure()
-    s.run()
-    assert(os.path.exists(os.path.join(test_out_dir, "2016-01-16","network_centralities.json")))
+    from pyveg.src.process_satellite_data import get_vegetation, get_weather
+    print('Warning: this test is expected to take a while to run...')
+    nc_results = get_vegetation(test_out_dir, data_collections['Copernicus'], coordinates, date_range, n_sub_images=1)
+    assert( len(nc_results) != 0 )
     shutil.rmtree(test_out_dir, ignore_errors=True)
-
-
 
 
 @unittest.skipIf(os.environ.get('TRAVIS') == 'true','Skipping this test on Travis CI.')
 def test_get_rainfall():
-    from pyveg.src.download_modules import WeatherDownloader
-    s = Sequence("weather")
-    s.set_config(data_collections["ERA5"])
-    s.output_dir = test_out_dir
-    s.coords = coordinates
-    s.date_range = date_range
-    s.time_per_point = time_per_point
-    s += WeatherDownloader()
-    s += WeatherImageToJSON()
-    s.configure()
-    s.run()
-    assert(os.path.exists(os.path.join(test_out_dir, "RESULTS", "weather_data.json")))
+    from pyveg.src.process_satellite_data import get_vegetation, get_weather
+    print('Warning: this test is expected to take a while to run...')
+    result = get_weather(test_out_dir, data_collections['NOAA'], coordinates, date_range)
+    assert (len(result.items())!=0)
     shutil.rmtree(test_out_dir, ignore_errors=True)
