+++ setup.py
@@ -19,7 +19,6 @@ setup(
         "pyveg_calc_EC=pyveg.scripts.calc_euler_characteristic:main",
         "pyveg_gen_pattern=pyveg.scripts.generate_pattern:main",
         "pyveg_gee_download=pyveg.scripts.download_gee_data:main",
-        "pyveg_gee_analysis=pyveg.scripts.analyse_gee_data:main",
-        "pyveg_run_pipeline=pyveg.scripts.run_pyveg_pipeline:main"
+        "pyveg_gee_analysis=pyveg.scripts.analyse_gee_data:main"
     ]},
 )
